% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdm-package.R
\docType{package}
\name{wdm-package}
\alias{wdm}
\alias{wdm-package}
\title{wdm: Weighted Dependence Measures}
\description{
Provides efficient implementations of weighted dependence measures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tnagler/wdm}
  \item Report bugs at \url{https://github.com/tnagler/wdm/issues}
}

}
\author{
\strong{Maintainer}: Thomas Nagler \email{mail@tnagler.com}

}
\keyword{internal}
