% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdm-package.R
\docType{package}
\name{wdm-package}
\alias{wdm-package}
\title{\packageTitle{wdm}}
\description{
\packageDescription{wdm}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{VineCopula}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tnagler/wdm-r}
  \item Report bugs at \url{https://github.com/tnagler/wdm-r/issues}
}

}
\author{
\strong{Maintainer}: Thomas Nagler \email{mail@tnagler.com}

}
